M2-6242: Unit tests for activity entity mapToActivity function
